Move Session class under dedicated namespace
